requirements.txt
streamlit==1.33.0
openai-whisper==20231111
tqdm==4.66.2
ffmpeg-python==0.2.0